SF Fatality Analysis from 2015-2017
When filtering records for only 2015-2017, there are only ~116 records that are specific to San Francisco County. I am not sure if that is enough for a significant analysis. However, I can go further back towards 2010 to gather a larger sample size. There are minor differences in the fields that can be rectified in R.