New rmd file with more suitable dataframe
This rmd file reflects the updated direction in which the project is going. The dataframe includes fatal and non-fatal traffic incidents in San Francisco, excluding those reported by the California Highway Patrol (CHP) on interstate highways. Further tidying will be done in coding string values to integer values using the guide that was provided by the CHP.